Fix soundless builds and car minigame crash (#292)

+++ src/anniversary/sf33rd/Source/Game/sc_sub.c
@@ -1939,7 +1939,7 @@ void sc_ram_to_vram(s8 sc_num) {
 }
 
 void sc_ram_to_vram_opc(s8 sc_num, s8 x, s8 y, u16 atr) {
-    u32* sc_tbl_ptr;
+    uintptr_t* sc_tbl_ptr;
     u8* sc_pos_ptr;
     u8* sc_uv_ptr;
     u16 loop;
@@ -1950,7 +1950,7 @@ void sc_ram_to_vram_opc(s8 sc_num, s8 x, s8 y, u16 atr) {
     }
 
     ppgSetupCurrentDataList(&ppgScrList);
-    sc_tbl_ptr = (u32*)sc_ram_vram_tbl[sc_num];
+    sc_tbl_ptr = (uintptr_t*)sc_ram_vram_tbl[sc_num];
     sc_pos_ptr = (u8*)*sc_tbl_ptr;
     *sc_tbl_ptr++;
     sc_uv_ptr = (u8*)*sc_tbl_ptr;
